utils: Move to single kotlin file
