+++ app/src/main/java/com/example/notificationapp/data/network/api/RetrofitAccessObject.kt
@@ -1,6 +1,6 @@
 package com.example.notificationapp.data.network.api
 
-import com.example.notificationapp.utils.Constants
+import com.example.notificationapp.Constants
 import com.google.gson.GsonBuilder
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
