+++ app/src/main/java/com/example/notificationapp/view/activities/LoginActivity.kt
@@ -5,20 +5,11 @@ import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Bundle
 import android.util.Log
-import android.view.View
-import android.widget.Button
-import android.widget.TextView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.example.notificationapp.R
+import com.example.notificationapp.Constants
 import com.example.notificationapp.databinding.ActivityLoginBinding
-import com.example.notificationapp.utils.Constants
-import com.google.android.gms.tasks.Task
-import com.google.android.material.textfield.TextInputEditText
-import com.google.firebase.auth.AuthResult
 import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.GetTokenResult
-import java.util.*
 
 class LoginActivity : AppCompatActivity() {
